siderian-fastapi-app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic 
pyodbc
mysql-connector-python
apache-airflow
